+++ packages/bridge-controller/src/bridge-controller-method-action-types.ts
@@ -15,6 +15,11 @@ export type BridgeControllerFetchQuotesAction = {
   handler: BridgeController['fetchQuotes'];
 };
 
+export type BridgeControllerUpdateBatchSellTradesAction = {
+  type: `BridgeController:updateBatchSellTrades`;
+  handler: BridgeController['updateBatchSellTrades'];
+};
+
 export type BridgeControllerStopPollingForQuotesAction = {
   type: `BridgeController:stopPollingForQuotes`;
   handler: BridgeController['stopPollingForQuotes'];
@@ -23,6 +28,11 @@ export type BridgeControllerStopPollingForQuotesAction = {
 export type BridgeControllerSetLocationAction = {
   type: `BridgeController:setLocation`;
   handler: BridgeController['setLocation'];
+};
+
+export type BridgeControllerSetInputPrimaryDenominationAction = {
+  type: `BridgeController:setInputPrimaryDenomination`;
+  handler: BridgeController['setInputPrimaryDenomination'];
 };
 
 export type BridgeControllerResetStateAction = {
@@ -40,20 +50,16 @@ export type BridgeControllerTrackUnifiedSwapBridgeEventAction = {
   handler: BridgeController['trackUnifiedSwapBridgeEvent'];
 };
 
-export type BridgeControllerUpdateBatchSellTradesAction = {
-  type: `BridgeController:updateBatchSellTrades`;
-  handler: BridgeController['updateBatchSellTrades'];
-};
-
 /**
  * Union of all BridgeController action types.
  */
 export type BridgeControllerMethodActions =
   | BridgeControllerUpdateBridgeQuoteRequestParamsAction
   | BridgeControllerFetchQuotesAction
+  | BridgeControllerUpdateBatchSellTradesAction
   | BridgeControllerStopPollingForQuotesAction
   | BridgeControllerSetLocationAction
+  | BridgeControllerSetInputPrimaryDenominationAction
   | BridgeControllerResetStateAction
   | BridgeControllerSetChainIntervalLengthAction
-  | BridgeControllerTrackUnifiedSwapBridgeEventAction
-  | BridgeControllerUpdateBatchSellTradesAction;
+  | BridgeControllerTrackUnifiedSwapBridgeEventAction;